refactor: move admin consultation queries to db layer and use server component
Extract raw Prisma queries from API routes into centralized functions
in src/lib/db/consultations.ts (getConsultationsForAdmin, createConsultation,
updateConsultation, deleteConsultation, getAdminCityOptions).

Split the admin consultations page into a server component (auth + initial
data fetch) and a client component (UI), matching the pattern used by
other admin pages like offers and meetings.
